Перенес все в один интерактор
