--- app/src/main/java/ru/practicum/android/diploma/domain/CountryInteractor.kt
@@ -1,9 +1,0 @@
-package ru.practicum.android.diploma.domain
-
-import kotlinx.coroutines.flow.Flow
-import ru.practicum.android.diploma.domain.models.Country
-import ru.practicum.android.diploma.util.ResponseData
-
-interface CountryInteractor {
-    suspend fun getCountries(): Flow<ResponseData<List<Country>>>
-}
