--- app/src/main/java/ru/practicum/android/diploma/domain/IndustryInteractor.kt
@@ -1,9 +1,0 @@
-package ru.practicum.android.diploma.domain
-
-import kotlinx.coroutines.flow.Flow
-import ru.practicum.android.diploma.domain.models.Industries
-import ru.practicum.android.diploma.util.ResponseData
-
-interface IndustryInteractor {
-    suspend fun getIndustries(): Flow<ResponseData<List<Industries>>>
-}
